contatct related CRUD operation upadated

+++ mycontacys_backend/controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 //@desc get all contacts
 //@route /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async(req,res)=>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id : req.user.id}); 
     res.status(200).json(contacts);
 })
 
 //@desc get specific contact
 //@route /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async(req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -23,7 +23,7 @@ const getContact = asyncHandler(async(req,res)=>{
 
 //@desc post contacts
 //@route /api/contacts
-//@access public
+//@access private
 const postContacts = asyncHandler(async(req,res)=>{
     console.log("body request is :", req.body);
     const {name,email,phone} = req.body;
@@ -32,19 +32,26 @@ const postContacts = asyncHandler(async(req,res)=>{
         throw new Error("all fields are mandatory !");
     }
     const contact = await Contact.create({
-      name,email,phone
+      name,
+      email,
+      phone,
+      user_id : req.user.id
     });
     res.status(200).json(contact);
 })
 
 //@desc update contact
 //@route /api/contacts/:id
-//@access public
+//@access private
 const updateContacts = asyncHandler(async(req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
       res.status(404);
       throw new Error ("contact not found");
+    }
+    if(contact.user_id.toString() !== req.user.id ){
+      res.status(403);
+      throw new Error ("Unauthorized User can't update a contact");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -56,7 +63,7 @@ const updateContacts = asyncHandler(async(req,res)=>{
 
 //@desc delete contact
 //@route /api/contacts/:id
-//@access public
+//@access private
 const deleteContacts = asyncHandler(async (req, res) => {
     const id = req.params.id; 
     try {
@@ -65,6 +72,10 @@ const deleteContacts = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
       } 
+      if (contact.user_id.toString() !== req.user.id ){
+        res.status(403);
+        throw new Error ("Unauthorized User is can't delete a contact");
+      }
       await Contact.deleteOne({ _id: id }); 
       res.status(200).json({ message: "Contact deleted successfully" });
     } catch (error) {
